crates: add new experimental composefs-fuse crate
This allows us to mount a FUSE filesystem from the in-memory filesystem
tree.  That's useful because using erofs requires CAP_SYS_ADMIN in the
root namespace, which we often won't have.

This is not particularly complete.  It's a sort of minimal-effort first
implementation, but already basically works.

It's also not particularly efficient. In particular, it doesn't use
splice or passthrough fds.  Passthrough fds also currently require root
(although the FUSE developers are working on easing that restriction)
and although the Rust binding layer for FUSE recently gained passthrough
fd support, it hasn't been related yet.  Splice support is also missing
from the Rust binding layer.

Signed-off-by: Allison Karlitskaya <allison.karlitskaya@redhat.com>