composefs: fix feature flags of once_cell
We use once_cell because https://github.com/rust-lang/rust/issues/109737

once_cell has a large number of potential locking backends for `no_std`
environments.  We were randomly using `critical-section`, which itself
requires specifying a backend, which we weren't doing.  So far, we've
been protected by this by the fact that we also have once_cell as a
transient dependency (via tempfile).

Fix this to just use `std` so that we get locking via the stdlib.

Signed-off-by: Allison Karlitskaya <allison.karlitskaya@redhat.com>