repository: add insecure mode
If a repository is set to insecure, it will files that have fs-verity
disabled. Specifically it will continue as normal on
`MeasureVerityError:VerityMissing`.

A repository will still run with fs-verity enabled if it is able to
measure the verity of an image when it is opened, even if the repository
was set to insecure.

Signed-off-by: Sanne Raymaekers <sanne.raymaekers@gmail.com>