mount: clean up mount APIs
Change the Repository::mount() API to return the mounted filesystem as
an fd rather than taking the mountpoint as an argument.  Create a new
mount_at() API to replace the old one, replacing the canicalize() and
mount_at() calls that used to be in mount_composefs_at(), which we
remove.

Update the various users.

Making this change lets us simplify the logic in composefs-setup-root:
it no longer has to manually open the image in order to perform the
fsmount operation: it can use the new API on the repository.

This allows us to make Repository::open_image() private, so do that too.

Co-authored-by: Sanne Raymaekers <sanne.raymaekers@gmail.com>
Signed-off-by: Allison Karlitskaya <allison.karlitskaya@redhat.com>