fsverity: split `enable_verity` into multiple related functions
`enable_verity_raw` is effectively the same as the previous
`enable_verity`, it simply calls the underlying ioctl.

`enable_verity_with_retry` will attempt to enable verity an
opinionated three times with a 1ms delay between tries when
encountering ETXTBSY.  This is generally enough to work around the
delay with CLOEXEC in conjunction with fork()->exec().

`enable_verity_maybe_copy` will take ownership of the file descriptor
and attempt to enable verity using `enable_verity_with_retry`.  If
successful, it returns `VerityFd::Orig(fd)` to designate that the file
descriptor is the same one provided by the caller.  If unsuccessful,
it will instead copy the data to a new O_TMPFILE and return that via
`VerityFd::Copy(fd)`.  It is up to the caller to reconcile how this
new file descriptor is to be persisted into the filesystem.

Resolves: https://github.com/containers/composefs-rs/issues/106
Signed-off-by: John Eckersberg <jeckersb@redhat.com>