examples/bls: improve Containerfiles caching
Run the "install packages into the base image" section as the first
thing that happens, before the cfsctl binary gets copied in.

This means that we have to regenerate the initramfs, which is duplicated
work, but it enables caching of the package installation, which takes
much longer and involves downloading updated packages (which drift with
time).  This helps a lot with iterative local testing, and the bls
images are the ones that I usually use for that.

Also enable package caching for Arch and Debian by adding the
appropriate bind mounts.  Use `apt-get` instead of CLI-API-unstable
`apt`.  Use the new `kernel-install add-all` verb where applicable.

We leave Ubuntu out of this change for now: I keep ending up with two
initramfs files generated and I'm not sure how to workaround the weird
use of dracut there.  We can circle back to this later.

Signed-off-by: Allison Karlitskaya <allison.karlitskaya@redhat.com>