examples: add test.thing workarounds
We want to start running images with test.thing, so add the workarounds
for improved ssh-vsock support.  These are no longer necessary in arch
and fedora-rawhide.

Unfortunately we can't put these in common/ because it's outside of the
build context.  Using an extra build context also seems not to work
because symlinks aren't copied properly unless it's from the primary
context.

Make a small fix to a comment in examples/uki/Containerfile that should
have been cleaned up as part of a4cbd3e143a1 ("Update approach to
handling boot resources").

Signed-off-by: Allison Karlitskaya <allison.karlitskaya@redhat.com>