Split off `mod generic_tree`
I copied this code into https://github.com/cgwalters/tarsmith/blob/main/src/tree.rs
because I wanted a virtual FS but that wasn't specific to fsverity.

This adds a `mod generic_tree` which abstracts over that - then
the existing `mod tree` just plugs in the core composefs strategy
of storing regular files as an enum that can be either inline
or external.

Signed-off-by: Colin Walters <walters@verbum.org>