oci: add workaround for containers/skopeo#2750
We're going to start referring to OCI images by their names starting
with `sha256:` (as podman names them in the `--iidfile`) soon.  Skopeo
doesn't like that, so add a workaround.

This will soon let us get rid of some of the hacking about we do in our
`examples/` build scripts.

Signed-off-by: Allison Karlitskaya <allison.karlitskaya@redhat.com>