repository: change our "sync" story
The `fdatasync()` per written object thing is an unmitigated performance
disaster and we need to get rid of it.  It forces the filesystem to
create thousands upon thousands of tiny commits.

I tried another approach where we would reopen and `fsync()` each object
file referred to from a splitstream *after* all of the files were
written, but before we wrote the splitstream data, but it was also quite
bad.

syncfs() is a really really dumb hammer, and it could get us into
trouble if other users on the system have massive outstanding amounts of
IO.  On the other hand, it works: it's almost instantaneous, and is a
massive performance improvement over what we were doing before.  Let's
just go with that for now.

Maybe some day filesystems will have a mechanism for this which isn't
horrible.

Signed-off-by: Allison Karlitskaya <allison.karlitskaya@redhat.com>